Added a new flag for summarizing detection events
And consolidate Instruction Pointer Probe into a single event type
for Register Probe (since IP is a register.)

When events are detected, Zerotect knows which raw/OS events were
used in justifying that detection. Until now, we attached
all justifying events in full fidelity to every detected event.

This is not all that expensive in memory (they're just Arc pointers
being copied and held.), but gets prohibitively expensive
and redundant when sending over a network. The serialization cost
alone is heavy, and the numerous bytes on the network, more so.

This change allows three Justification modes:
1. Full Events (i.e. how it works today)
2. Summary (this is the new default): The register being probed
   is stored, and all values it takes in serial order.
   This allows us to see if the value distances and ordering
   was sufficient justification.
3. None: In this mode, merely the count of justifying events
   is provided.